Large refactoring to support testing on JVM and JS and on packages org.threeten.bp and java.time
