+++ jvm/src/test/scala/org/threeten/bp/format/TestReducedParser.scala
@@ -37,16 +37,17 @@ import org.threeten.bp.temporal.ChronoField.DAY_OF_YEAR
 import org.threeten.bp.temporal.ChronoField.YEAR
 import org.testng.annotations.DataProvider
 import org.testng.annotations.Test
+import org.threeten.bp.format.internal.TTBPDateTimeFormatterBuilder
 import org.threeten.bp.temporal.TemporalField
 import org.threeten.bp.temporal.TemporalQueries
 
 /** Test ReducedPrinterParser. */
 @Test class TestReducedParser extends AbstractTestPrinterParser {
   @DataProvider(name = "error") private[format] def data_error: Array[Array[Any]] = {
-    Array[Array[Any]](Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "12", -1, classOf[IndexOutOfBoundsException]), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "12", 3, classOf[IndexOutOfBoundsException]))
+    Array[Array[Any]](Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "12", -1, classOf[IndexOutOfBoundsException]), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "12", 3, classOf[IndexOutOfBoundsException]))
   }
 
-  @Test(dataProvider = "error") def test_parse_error(pp: DateTimeFormatterBuilder.ReducedPrinterParser, text: String, pos: Int, expected: Class[_]): Unit = {
+  @Test(dataProvider = "error") def test_parse_error(pp: TTBPDateTimeFormatterBuilder.ReducedPrinterParser, text: String, pos: Int, expected: Class[_]): Unit = {
     try pp.parse(parseContext, text, pos)
     catch {
       case ex: RuntimeException =>
@@ -58,27 +59,27 @@ import org.threeten.bp.temporal.TemporalQueries
 
   @throws(classOf[Exception])
   def test_parse_fieldRangeIgnored(): Unit = {
-    val pp: DateTimeFormatterBuilder.ReducedPrinterParser = new DateTimeFormatterBuilder.ReducedPrinterParser(DAY_OF_YEAR, 3, 3, 10, null)
+    val pp: TTBPDateTimeFormatterBuilder.ReducedPrinterParser = new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(DAY_OF_YEAR, 3, 3, 10, null)
     val newPos: Int = pp.parse(parseContext, "456", 0)
     assertEquals(newPos, 3)
     assertParsed(DAY_OF_YEAR, 456L)
   }
 
   @DataProvider(name = "Parse") private[format] def provider_parse: Array[Array[Any]] = {
-    Array[Array[Any]](Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "-0", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "Xxx12Xxx", 3, 5, 2012), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "12345", 0, 2, 2012), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "12-45", 0, 2, 2012), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "0", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "1", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "1", 1, ~1, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "1-2", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "9", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "A0", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "0A", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "  1", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "-1", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "-10", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "0", 0, 1, 2010), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "9", 0, 1, 2019), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "10", 0, 1, 2011), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "0", 0, 1, 2010), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "4", 0, 1, 2014), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "5", 0, 1, 2005), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "9", 0, 1, 2009), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "10", 0, 1, 2011), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "00", 0, 2, 2100), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "09", 0, 2, 2109), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "10", 0, 2, 2010), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "99", 0, 2, 2099), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "100", 0, 2, 2010), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "05", 0, 2, -2005), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "00", 0, 2, -2000), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "99", 0, 2, -1999), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "06", 0, 2, -1906), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "100", 0, 2, -1910))
+    Array[Array[Any]](Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "-0", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "Xxx12Xxx", 3, 5, 2012), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "12345", 0, 2, 2012), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "12-45", 0, 2, 2012), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "0", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "1", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "1", 1, ~1, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "1-2", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "9", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "A0", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "0A", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "  1", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "-1", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "-10", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "0", 0, 1, 2010), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "9", 0, 1, 2019), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "10", 0, 1, 2011), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "0", 0, 1, 2010), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "4", 0, 1, 2014), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "5", 0, 1, 2005), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "9", 0, 1, 2009), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "10", 0, 1, 2011), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "00", 0, 2, 2100), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "09", 0, 2, 2109), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "10", 0, 2, 2010), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "99", 0, 2, 2099), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "100", 0, 2, 2010), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "05", 0, 2, -2005), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "00", 0, 2, -2000), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "99", 0, 2, -1999), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "06", 0, 2, -1906), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "100", 0, 2, -1910))
   }
 
-  @Test(dataProvider = "Parse") def test_parse(pp: DateTimeFormatterBuilder.ReducedPrinterParser, input: String, pos: Int, parseLen: Int, parseVal: Integer): Unit = {
+  @Test(dataProvider = "Parse") def test_parse(pp: TTBPDateTimeFormatterBuilder.ReducedPrinterParser, input: String, pos: Int, parseLen: Int, parseVal: Integer): Unit = {
     val newPos: Int = pp.parse(parseContext, input, pos)
     assertEquals(newPos, parseLen)
     assertParsed(YEAR, if (parseVal != null) parseVal.toLong else null)
   }
 
   @DataProvider(name = "ParseLenient") private[format] def provider_parseLenient: Array[Array[Any]] = {
-    Array[Array[Any]](Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "-0", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "Xxx12Xxx", 3, 5, 2012), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "12345", 0, 5, 12345), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "12-45", 0, 2, 2012), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "0", 0, 1, 0), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "1", 0, 1, 1), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "1", 1, ~1, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "1-2", 0, 1, 1), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "9", 0, 1, 9), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "A0", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "0A", 0, 1, 0), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "  1", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "-1", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "-10", 0, ~0, null), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "0", 0, 1, 2010), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "9", 0, 1, 2019), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "10", 0, 2, 10), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "0", 0, 1, 2010), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "4", 0, 1, 2014), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "5", 0, 1, 2005), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "9", 0, 1, 2009), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "10", 0, 2, 10), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "00", 0, 2, 2100), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "09", 0, 2, 2109), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "10", 0, 2, 2010), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "99", 0, 2, 2099), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "100", 0, 3, 100), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "05", 0, 2, -2005), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "00", 0, 2, -2000), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "99", 0, 2, -1999), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "06", 0, 2, -1906), Array(new DateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "100", 0, 3, 100))
+    Array[Array[Any]](Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "-0", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "Xxx12Xxx", 3, 5, 2012), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "12345", 0, 5, 12345), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "12-45", 0, 2, 2012), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "0", 0, 1, 0), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "1", 0, 1, 1), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "1", 1, ~1, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "1-2", 0, 1, 1), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "9", 0, 1, 9), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "A0", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "0A", 0, 1, 0), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "  1", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "-1", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "-10", 0, ~0, null), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "0", 0, 1, 2010), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "9", 0, 1, 2019), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2010, null), "10", 0, 2, 10), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "0", 0, 1, 2010), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "4", 0, 1, 2014), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "5", 0, 1, 2005), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "9", 0, 1, 2009), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 1, 1, 2005, null), "10", 0, 2, 10), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "00", 0, 2, 2100), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "09", 0, 2, 2109), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "10", 0, 2, 2010), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "99", 0, 2, 2099), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, 2010, null), "100", 0, 3, 100), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "05", 0, 2, -2005), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "00", 0, 2, -2000), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "99", 0, 2, -1999), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "06", 0, 2, -1906), Array(new TTBPDateTimeFormatterBuilder.ReducedPrinterParser(YEAR, 2, 2, -2005, null), "100", 0, 3, 100))
   }
 
-  @Test(dataProvider = "ParseLenient") def test_parseLenient(pp: DateTimeFormatterBuilder.ReducedPrinterParser, input: String, pos: Int, parseLen: Int, parseVal: Integer): Unit = {
+  @Test(dataProvider = "ParseLenient") def test_parseLenient(pp: TTBPDateTimeFormatterBuilder.ReducedPrinterParser, input: String, pos: Int, parseLen: Int, parseVal: Integer): Unit = {
     parseContext.setStrict(false)
     val newPos: Int = pp.parse(parseContext, input, pos)
     assertEquals(newPos, parseLen)
