+++ jvm/src/test/scala/org/threeten/bp/format/TestTextParser.scala
@@ -37,10 +37,12 @@ import org.threeten.bp.temporal.ChronoField.DAY_OF_MONTH
 import org.threeten.bp.temporal.ChronoField.DAY_OF_WEEK
 import org.threeten.bp.temporal.ChronoField.MONTH_OF_YEAR
 import java.util.Locale
+
 import org.testng.annotations.DataProvider
 import org.testng.annotations.Test
 import org.threeten.bp.temporal.TemporalField
 import org.threeten.bp.temporal.TemporalQueries
+import org.threeten.bp.format.internal.{TTBPDateTimeFormatterBuilder, TTBPDateTimeParseContext}
 
 /** Test TextPrinterParser. */
 object TestTextParser {
@@ -49,10 +51,10 @@ object TestTextParser {
 
 @Test class TestTextParser extends AbstractTestPrinterParser {
   @DataProvider(name = "error") private[format] def data_error: Array[Array[Any]] = {
-    Array[Array[Any]](Array(new DateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL, TestTextParser.PROVIDER), "Monday", -1, classOf[IndexOutOfBoundsException]), Array(new DateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL, TestTextParser.PROVIDER), "Monday", 7, classOf[IndexOutOfBoundsException]))
+    Array[Array[Any]](Array(new TTBPDateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL, TestTextParser.PROVIDER), "Monday", -1, classOf[IndexOutOfBoundsException]), Array(new TTBPDateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL, TestTextParser.PROVIDER), "Monday", 7, classOf[IndexOutOfBoundsException]))
   }
 
-  @Test(dataProvider = "error") def test_parse_error(pp: DateTimeFormatterBuilder.TextPrinterParser, text: String, pos: Int, expected: Class[_]): Unit = {
+  @Test(dataProvider = "error") def test_parse_error(pp: TTBPDateTimeFormatterBuilder.TextPrinterParser, text: String, pos: Int, expected: Class[_]): Unit = {
     try {
       pp.parse(parseContext, text, pos)
     }
@@ -66,7 +68,7 @@ object TestTextParser {
 
   @throws(classOf[Exception])
   def test_parse_midStr(): Unit = {
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "XxxMondayXxx", 3)
     assertEquals(newPos, 9)
     assertParsed(parseContext, DAY_OF_WEEK, 1L)
@@ -74,7 +76,7 @@ object TestTextParser {
 
   @throws(classOf[Exception])
   def test_parse_remainderIgnored(): Unit = {
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.SHORT, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.SHORT, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "Wednesday", 0)
     assertEquals(newPos, 3)
     assertParsed(parseContext, DAY_OF_WEEK, 3L)
@@ -82,7 +84,7 @@ object TestTextParser {
 
   @throws(classOf[Exception])
   def test_parse_noMatch1(): Unit = {
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "Munday", 0)
     assertEquals(newPos, ~0)
     assertEquals(parseContext.toParsed.query(TemporalQueries.chronology), null)
@@ -91,7 +93,7 @@ object TestTextParser {
 
   @throws(classOf[Exception])
   def test_parse_noMatch2(): Unit = {
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "Monday", 3)
     assertEquals(newPos, ~3)
     assertEquals(parseContext.toParsed.query(TemporalQueries.chronology), null)
@@ -100,7 +102,7 @@ object TestTextParser {
 
   @throws(classOf[Exception])
   def test_parse_noMatch_atEnd(): Unit = {
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(DAY_OF_WEEK, TextStyle.FULL, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "Monday", 6)
     assertEquals(newPos, ~6)
     assertEquals(parseContext.toParsed.query(TemporalQueries.chronology), null)
@@ -118,7 +120,7 @@ object TestTextParser {
   @Test(dataProvider = "parseText")
   @throws(classOf[Exception])
   def test_parseText(field: TemporalField, style: TextStyle, value: Int, input: String): Unit = {
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(field, style, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(field, style, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, input, 0)
     assertEquals(newPos, input.length)
     assertParsed(parseContext, field, value.toLong)
@@ -127,7 +129,7 @@ object TestTextParser {
   @Test(dataProvider = "parseNumber")
   @throws(classOf[Exception])
   def test_parseNumber(field: TemporalField, style: TextStyle, value: Int, input: String): Unit = {
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(field, style, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(field, style, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, input, 0)
     assertEquals(newPos, input.length)
     assertParsed(parseContext, field, value.toLong)
@@ -137,7 +139,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_strict_caseSensitive_parseUpper(field: TemporalField, style: TextStyle, value: Int, input: String): Unit = {
     parseContext.setCaseSensitive(true)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(field, style, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(field, style, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, input.toUpperCase, 0)
     assertEquals(newPos, ~0)
     assertEquals(parseContext.toParsed.query(TemporalQueries.chronology), null)
@@ -148,7 +150,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_strict_caseInsensitive_parseUpper(field: TemporalField, style: TextStyle, value: Int, input: String): Unit = {
     parseContext.setCaseSensitive(false)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(field, style, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(field, style, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, input.toUpperCase, 0)
     assertEquals(newPos, input.length)
     assertParsed(parseContext, field, value.toLong)
@@ -158,7 +160,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_strict_caseSensitive_parseLower(field: TemporalField, style: TextStyle, value: Int, input: String): Unit = {
     parseContext.setCaseSensitive(true)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(field, style, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(field, style, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, input.toLowerCase, 0)
     assertEquals(newPos, ~0)
     assertEquals(parseContext.toParsed.query(TemporalQueries.chronology), null)
@@ -169,7 +171,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_strict_caseInsensitive_parseLower(field: TemporalField, style: TextStyle, value: Int, input: String): Unit = {
     parseContext.setCaseSensitive(false)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(field, style, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(field, style, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, input.toLowerCase, 0)
     assertEquals(newPos, input.length)
     assertParsed(parseContext, field, value.toLong)
@@ -178,7 +180,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_full_strict_full_match(): Unit = {
     parseContext.setStrict(true)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "January", 0)
     assertEquals(newPos, 7)
     assertParsed(parseContext, MONTH_OF_YEAR, 1L)
@@ -187,7 +189,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_full_strict_short_noMatch(): Unit = {
     parseContext.setStrict(true)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "Janua", 0)
     assertEquals(newPos, ~0)
     assertEquals(parseContext.toParsed.query(TemporalQueries.chronology), null)
@@ -197,7 +199,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_full_strict_number_noMatch(): Unit = {
     parseContext.setStrict(true)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "1", 0)
     assertEquals(newPos, ~0)
     assertEquals(parseContext.toParsed.query(TemporalQueries.chronology), null)
@@ -207,7 +209,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_short_strict_full_match(): Unit = {
     parseContext.setStrict(true)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "January", 0)
     assertEquals(newPos, 3)
     assertParsed(parseContext, MONTH_OF_YEAR, 1L)
@@ -216,7 +218,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_short_strict_short_match(): Unit = {
     parseContext.setStrict(true)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "Janua", 0)
     assertEquals(newPos, 3)
     assertParsed(parseContext, MONTH_OF_YEAR, 1L)
@@ -225,7 +227,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_short_strict_number_noMatch(): Unit = {
     parseContext.setStrict(true)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "1", 0)
     assertEquals(newPos, ~0)
     assertEquals(parseContext.toParsed.query(TemporalQueries.chronology), null)
@@ -236,7 +238,7 @@ object TestTextParser {
   def test_parse_french_short_strict_full_noMatch(): Unit = {
     parseContext.setLocale(Locale.FRENCH)
     parseContext.setStrict(true)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "janvier", 0)
     assertEquals(newPos, ~0)
     assertEquals(parseContext.toParsed.query(TemporalQueries.chronology), null)
@@ -247,7 +249,7 @@ object TestTextParser {
   def test_parse_french_short_strict_short_match(): Unit = {
     parseContext.setLocale(Locale.FRENCH)
     parseContext.setStrict(true)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "janv.", 0)
     assertEquals(newPos, 5)
     assertParsed(parseContext, MONTH_OF_YEAR, 1L)
@@ -256,7 +258,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_full_lenient_full_match(): Unit = {
     parseContext.setStrict(false)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "January", 0)
     assertEquals(newPos, 7)
     assertParsed(parseContext, MONTH_OF_YEAR, 1L)
@@ -265,7 +267,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_full_lenient_short_match(): Unit = {
     parseContext.setStrict(false)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "Janua", 0)
     assertEquals(newPos, 3)
     assertParsed(parseContext, MONTH_OF_YEAR, 1L)
@@ -274,7 +276,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_full_lenient_number_match(): Unit = {
     parseContext.setStrict(false)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.FULL, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "1", 0)
     assertEquals(newPos, 1)
     assertParsed(parseContext, MONTH_OF_YEAR, 1L)
@@ -283,7 +285,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_short_lenient_full_match(): Unit = {
     parseContext.setStrict(false)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "January", 0)
     assertEquals(newPos, 7)
     assertParsed(parseContext, MONTH_OF_YEAR, 1L)
@@ -292,7 +294,7 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_short_lenient_short_match(): Unit = {
     parseContext.setStrict(false)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "Janua", 0)
     assertEquals(newPos, 3)
     assertParsed(parseContext, MONTH_OF_YEAR, 1L)
@@ -301,13 +303,13 @@ object TestTextParser {
   @throws(classOf[Exception])
   def test_parse_short_lenient_number_match(): Unit = {
     parseContext.setStrict(false)
-    val pp: DateTimeFormatterBuilder.TextPrinterParser = new DateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
+    val pp: TTBPDateTimeFormatterBuilder.TextPrinterParser = new TTBPDateTimeFormatterBuilder.TextPrinterParser(MONTH_OF_YEAR, TextStyle.SHORT, TestTextParser.PROVIDER)
     val newPos: Int = pp.parse(parseContext, "1", 0)
     assertEquals(newPos, 1)
     assertParsed(parseContext, MONTH_OF_YEAR, 1L)
   }
 
-  private def assertParsed(context: DateTimeParseContext, field: TemporalField, value: java.lang.Long): Unit = {
+  private def assertParsed(context: TTBPDateTimeParseContext, field: TemporalField, value: java.lang.Long): Unit = {
     if (value == null) {
       assertEquals(context.getParsed(field), null)
     }
