+++ shared/src/main/scala/org/threeten/bp/chrono/Era.scala
@@ -32,9 +32,18 @@
 package org.threeten.bp.chrono
 
 import java.util.Locale
-import org.threeten.bp.format.{DateTimeFormatterBuilder, TextStyle}
-import org.threeten.bp.temporal.ChronoField._
-import org.threeten.bp.temporal._
+import org.threeten.bp.format.DateTimeFormatterBuilder
+import org.threeten.bp.format.TextStyle
+import org.threeten.bp.temporal.Temporal
+import org.threeten.bp.temporal.TemporalQueries
+import org.threeten.bp.temporal.TemporalQuery
+import org.threeten.bp.temporal.TemporalAccessor
+import org.threeten.bp.temporal.TemporalAdjuster
+import org.threeten.bp.temporal.TemporalField
+import org.threeten.bp.temporal.ChronoField
+import org.threeten.bp.temporal.ChronoField.ERA
+import org.threeten.bp.temporal.ChronoUnit
+import org.threeten.bp.temporal.UnsupportedTemporalTypeException
 
 /** An era of the time-line.
   *
