+++ shared/src/main/scala/org/threeten/bp/format/internal/TTBPDateTimeParseContext.scala
@@ -29,23 +29,33 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.threeten.bp.format
+package org.threeten.bp.format.internal
 
-import java.util.{Objects, Locale}
+import java.util.{Locale, Objects}
+
 import org.threeten.bp.Period
 import org.threeten.bp.ZoneId
 import org.threeten.bp.chrono.Chronology
 import org.threeten.bp.chrono.IsoChronology
-import org.threeten.bp.temporal._
+import org.threeten.bp.temporal.TemporalField
+import org.threeten.bp.temporal.TemporalAccessor
+import org.threeten.bp.temporal.UnsupportedTemporalTypeException
+import org.threeten.bp.temporal.TemporalQuery
+import org.threeten.bp.temporal.TemporalQueries
+import org.threeten.bp.format.DateTimeBuilder
+import org.threeten.bp.format.DecimalStyle
+import org.threeten.bp.format.DateTimeFormatterBuilder
+import org.threeten.bp.format.DateTimeFormatter
+import org.threeten.bp.format.internal.TTBPDateTimeFormatterBuilder.ReducedPrinterParser
 
-private[format] object DateTimeParseContext {
+ object TTBPDateTimeParseContext {
   /** Compares two characters ignoring case.
     *
     * @param c1  the first
     * @param c2  the second
     * @return true if equal
     */
-  private[format] def charEqualsIgnoreCase(c1: Char, c2: Char): Boolean =
+   def charEqualsIgnoreCase(c1: Char, c2: Char): Boolean =
     c1 == c2 || Character.toUpperCase(c1) == Character.toUpperCase(c2) || Character.toLowerCase(c1) == Character.toLowerCase(c2)
 }
 
@@ -63,16 +73,16 @@ private[format] object DateTimeParseContext {
   * Usage of the class is thread-safe within standard parsing as a new instance of this class
   * is automatically created for each parse and parsing is single-threaded
   */
-final class DateTimeParseContext private[format](private var locale: Locale,
-                                                 private var symbols: DecimalStyle,
-                                                 private var overrideChronology: Chronology,
-                                                 private var overrideZone: ZoneId,
-                                                 private var caseSensitive: Boolean = true,
-                                                 private var strict: Boolean = true) {
+final class TTBPDateTimeParseContext (private var locale: Locale,
+                                                     private var symbols: DecimalStyle,
+                                                     private var overrideChronology: Chronology,
+                                                     private var overrideZone: ZoneId,
+                                                     private var caseSensitive: Boolean = true,
+                                                     private var strict: Boolean = true) {
   /** The list of parsed data.
     */
-  private val parsed: java.util.ArrayList[DateTimeParseContext#Parsed] = {
-    val list = new java.util.ArrayList[DateTimeParseContext#Parsed]
+  private val parsed: java.util.ArrayList[TTBPDateTimeParseContext#Parsed] = {
+    val list = new java.util.ArrayList[TTBPDateTimeParseContext#Parsed]
     list.add(new Parsed)
     list
   }
@@ -81,21 +91,21 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     *
     * @param formatter  the formatter controlling the parse, not null
     */
-  private[format] def this(formatter: DateTimeFormatter) {
+   def this(formatter: DateTimeFormatter) {
     this(formatter.getLocale, formatter.getDecimalStyle, formatter.getChronology, formatter.getZone)
   }
 
-  private[format] def this(locale: Locale, symbols: DecimalStyle, chronology: Chronology) {
+   def this(locale: Locale, symbols: DecimalStyle, chronology: Chronology) {
     this(locale, symbols, chronology, null)
   }
 
-  private[format] def this(other: DateTimeParseContext) {
+   def this(other: TTBPDateTimeParseContext) {
     this(other.locale, other.symbols, other.overrideChronology, other.overrideZone, other.caseSensitive, other.strict)
   }
 
   /** Creates a copy of this context.
     */
-  private[format] def copy: DateTimeParseContext = new DateTimeParseContext(this)
+   def copy: TTBPDateTimeParseContext = new TTBPDateTimeParseContext(this)
 
   /** Gets the locale.
     *
@@ -104,7 +114,7 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     *
     * @return the locale, not null
     */
-  private[format] def getLocale: Locale = locale
+   def getLocale: Locale = locale
 
   /** Gets the formatting symbols.
     *
@@ -112,13 +122,13 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     *
     * @return the formatting symbols, not null
     */
-  private[format] def getSymbols: DecimalStyle = symbols
+   def getSymbols: DecimalStyle = symbols
 
   /** Gets the effective chronology during parsing.
     *
     * @return the effective parsing chronology, not null
     */
-  private[format] def getEffectiveChronology: Chronology = {
+   def getEffectiveChronology: Chronology = {
     var chrono: Chronology = currentParsed.chrono
     if (chrono == null) {
       chrono = overrideChronology
@@ -133,13 +143,13 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     *
     * @return true if parsing is case sensitive, false if case insensitive
     */
-  private[format] def isCaseSensitive: Boolean = caseSensitive
+   def isCaseSensitive: Boolean = caseSensitive
 
   /** Sets whether the parsing is case sensitive or not.
     *
     * @param caseSensitive  changes the parsing to be case sensitive or not from now on
     */
-  private[format] def setCaseSensitive(caseSensitive: Boolean): Unit = this.caseSensitive = caseSensitive
+   def setCaseSensitive(caseSensitive: Boolean): Unit = this.caseSensitive = caseSensitive
 
   /** Helper to compare two {@code CharSequence} instances.
     * This uses {@link #isCaseSensitive()}.
@@ -151,7 +161,7 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     * @param length  the length to check, valid
     * @return true if equal
     */
-  private[format] def subSequenceEquals(cs1: CharSequence, offset1: Int, cs2: CharSequence, offset2: Int, length: Int): Boolean = {
+   def subSequenceEquals(cs1: CharSequence, offset1: Int, cs2: CharSequence, offset2: Int, length: Int): Boolean = {
     if (offset1 + length > cs1.length || offset2 + length > cs2.length) {
       return false
     }
@@ -191,11 +201,11 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     * @param ch2  the second character
     * @return true if equal
     */
-  private[format] def charEquals(ch1: Char, ch2: Char): Boolean =
+   def charEquals(ch1: Char, ch2: Char): Boolean =
     if (isCaseSensitive)
       ch1 == ch2
     else
-      DateTimeParseContext.charEqualsIgnoreCase(ch1, ch2)
+      TTBPDateTimeParseContext.charEqualsIgnoreCase(ch1, ch2)
 
   /** Checks if parsing is strict.
     *
@@ -203,23 +213,23 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     *
     * @return true if parsing is strict, false if lenient
     */
-  private[format] def isStrict: Boolean = strict
+   def isStrict: Boolean = strict
 
   /** Sets whether parsing is strict or lenient.
     *
     * @param strict  changes the parsing to be strict or lenient from now on
     */
-  private[format] def setStrict(strict: Boolean): Unit = this.strict = strict
+   def setStrict(strict: Boolean): Unit = this.strict = strict
 
   /** Starts the parsing of an optional segment of the input.
     */
-  private[format] def startOptional(): Unit = parsed.add(currentParsed.copy)
+   def startOptional(): Unit = parsed.add(currentParsed.copy)
 
   /** Ends the parsing of an optional segment of the input.
     *
     * @param successful  whether the optional segment was successfully parsed
     */
-  private[format] def endOptional(successful: Boolean): Unit =
+   def endOptional(successful: Boolean): Unit =
     if (successful)
       parsed.remove(parsed.size - 2)
     else
@@ -229,7 +239,7 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     *
     * @return the current temporal objects, not null
     */
-  private def currentParsed: DateTimeParseContext#Parsed = parsed.get(parsed.size - 1)
+  private def currentParsed: TTBPDateTimeParseContext#Parsed = parsed.get(parsed.size - 1)
 
   /** Gets the first value that was parsed for the specified field.
     *
@@ -241,7 +251,7 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     * @param field  the field to query from the map, null returns null
     * @return the value mapped to the specified field, null if field was not parsed
     */
-  private[format] def getParsed(field: TemporalField): java.lang.Long = currentParsed.fieldValues.get(field)
+   def getParsed(field: TemporalField): java.lang.Long = currentParsed.fieldValues.get(field)
 
   /** Stores the parsed field.
     *
@@ -254,7 +264,7 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     * @param successPos  the position after the field being parsed
     * @return the new position
     */
-  private[format] def setParsedField(field: TemporalField, value: Long, errorPos: Int, successPos: Int): Int = {
+   def setParsedField(field: TemporalField, value: Long, errorPos: Int, successPos: Int): Int = {
     Objects.requireNonNull(field, "field")
     val old: java.lang.Long = currentParsed.fieldValues.put(field, value)
     if (old != null && old.longValue != value) ~errorPos else successPos
@@ -267,23 +277,23 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     *
     * @param chrono  the parsed chronology, not null
     */
-  private[format] def setParsed(chrono: Chronology): Unit = {
+   def setParsed(chrono: Chronology): Unit = {
     Objects.requireNonNull(chrono, "chrono")
-    val _currentParsed: DateTimeParseContext#Parsed = currentParsed
+    val _currentParsed: TTBPDateTimeParseContext#Parsed = currentParsed
     _currentParsed.chrono = chrono
     if (_currentParsed.callbacks != null) {
       val callbacks: java.util.Iterator[Array[AnyRef]] = new java.util.ArrayList[Array[AnyRef]](_currentParsed.callbacks).iterator
       _currentParsed.callbacks.clear()
       while (callbacks.hasNext) {
         val objects = callbacks.next()
-        val pp: DateTimeFormatterBuilder.ReducedPrinterParser = objects(0).asInstanceOf[DateTimeFormatterBuilder.ReducedPrinterParser]
+        val pp: ReducedPrinterParser = objects(0).asInstanceOf[ReducedPrinterParser]
         pp.setValue(this, objects(1).asInstanceOf[java.lang.Long], objects(2).asInstanceOf[Integer], objects(3).asInstanceOf[Integer])
       }
     }
   }
 
-  private[format] def addChronologyChangedParser(reducedPrinterParser: DateTimeFormatterBuilder.ReducedPrinterParser, value: Long, errorPos: Int, successPos: Int): Unit = {
-    val _currentParsed: DateTimeParseContext#Parsed = currentParsed
+   def addChronologyChangedParser(reducedPrinterParser: ReducedPrinterParser, value: Long, errorPos: Int, successPos: Int): Unit = {
+    val _currentParsed: TTBPDateTimeParseContext#Parsed = currentParsed
     if (_currentParsed.callbacks == null)
       _currentParsed.callbacks = new java.util.ArrayList[Array[AnyRef]](2)
     _currentParsed.callbacks.add(Array[AnyRef](reducedPrinterParser, value.asInstanceOf[AnyRef], errorPos.asInstanceOf[AnyRef], successPos.asInstanceOf[AnyRef]))
@@ -296,21 +306,21 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     *
     * @param zone  the parsed zone, not null
     */
-  private[format] def setParsed(zone: ZoneId): Unit = {
+   def setParsed(zone: ZoneId): Unit = {
     Objects.requireNonNull(zone, "zone")
     currentParsed.zone = zone
   }
 
   /** Stores the leap second.
     */
-  private[format] def setParsedLeapSecond(): Unit = currentParsed.leapSecond = true
+   def setParsedLeapSecond(): Unit = currentParsed.leapSecond = true
 
   /** Returns a {@code TemporalAccessor} that can be used to interpret
     * the results of the parse.
     *
     * @return an accessor with the results of the parse, not null
     */
-  private[format] def toParsed: DateTimeParseContext#Parsed = currentParsed
+   def toParsed: TTBPDateTimeParseContext#Parsed = currentParsed
 
   /** Returns a string version of the context for debugging.
     *
@@ -320,16 +330,16 @@ final class DateTimeParseContext private[format](private var locale: Locale,
 
   /** Temporary store of parsed data.
     */
-  private[format] final class Parsed private[format]() extends TemporalAccessor {
-    private[format] var chrono: Chronology = null
-    private[format] var zone: ZoneId = null
-    private[format] val fieldValues: java.util.Map[TemporalField, java.lang.Long] = new java.util.HashMap[TemporalField, java.lang.Long]
-    private[format] var leapSecond: Boolean = false
-    private[format] var excessDays: Period = Period.ZERO
-    private[format] var callbacks: java.util.List[Array[AnyRef]] = null
+   final class Parsed () extends TemporalAccessor {
+     var chrono: Chronology = null
+     var zone: ZoneId = null
+     val fieldValues: java.util.Map[TemporalField, java.lang.Long] = new java.util.HashMap[TemporalField, java.lang.Long]
+     var leapSecond: Boolean = false
+     var excessDays: Period = Period.ZERO
+     var callbacks: java.util.List[Array[AnyRef]] = null
 
-    protected[format] def copy: DateTimeParseContext#Parsed = {
-      val cloned: DateTimeParseContext#Parsed = new Parsed
+    protected[format] def copy: TTBPDateTimeParseContext#Parsed = {
+      val cloned: TTBPDateTimeParseContext#Parsed = new Parsed
       cloned.chrono = this.chrono
       cloned.zone = this.zone
       cloned.fieldValues.putAll(this.fieldValues)
@@ -373,7 +383,7 @@ final class DateTimeParseContext private[format](private var locale: Locale,
       *
       * @return a new builder with the results of the parse, not null
       */
-    private[format] def toBuilder: DateTimeBuilder = {
+     def toBuilder: DateTimeBuilder = {
       val builder: DateTimeBuilder = new DateTimeBuilder
       builder.fieldValues.putAll(fieldValues.asInstanceOf[java.util.Map[_ <: TemporalField, _ <: java.lang.Long]])
       builder.chrono = getEffectiveChronology
@@ -394,7 +404,7 @@ final class DateTimeParseContext private[format](private var locale: Locale,
     *
     * @param locale  the locale, not null
     */
-  private[format] def setLocale(locale: Locale): Unit = {
+   def setLocale(locale: Locale): Unit = {
     Objects.requireNonNull(locale, "locale")
     this.locale = locale
   }
