+++ shared/src/test/scala/org/threeten/bp/format/GenTestPrinterParser.scala
@@ -33,7 +33,7 @@ package org.threeten.bp.format
 
 import java.util.Locale
 
-import org.scalatest.{BeforeAndAfter, BeforeAndAfterEach, FunSuite}
+import org.scalatest.{BeforeAndAfterEach, FunSuite}
 import org.threeten.bp.LocalDateTime
 import org.threeten.bp.ZoneId
 import org.threeten.bp.ZonedDateTime
@@ -41,7 +41,8 @@ import org.threeten.bp.DateTimeException
 import org.threeten.bp.chrono.IsoChronology
 import org.threeten.bp.temporal.TemporalAccessor
 import org.threeten.bp.temporal.TemporalField
-
+import org.threeten.bp.format.internal.TTBPDateTimeParseContext
+import org.threeten.bp.format.internal.TTBPDateTimePrintContext
 
 /** Abstract PrinterParser test. */
 object GenTestPrinterParser {
@@ -52,16 +53,16 @@ object GenTestPrinterParser {
 }
 
 trait GenTestPrinterParser extends BeforeAndAfterEach { this: FunSuite =>
-  protected var printEmptyContext: DateTimePrintContext = null
-  protected var printContext: DateTimePrintContext = null
-  protected var parseContext: DateTimeParseContext = null
+  protected var printEmptyContext: TTBPDateTimePrintContext = null
+  protected var printContext: TTBPDateTimePrintContext = null
+  protected var parseContext: TTBPDateTimeParseContext = null
   protected var buf: java.lang.StringBuilder = null
 
   override def beforeEach() {
-    printEmptyContext = new DateTimePrintContext(GenTestPrinterParser.EMPTY, Locale.ENGLISH, DecimalStyle.STANDARD)
+    printEmptyContext = new TTBPDateTimePrintContext(GenTestPrinterParser.EMPTY, Locale.ENGLISH, DecimalStyle.STANDARD)
     val zdt: ZonedDateTime = LocalDateTime.of(2011, 6, 30, 12, 30, 40, 0).atZone(ZoneId.of("Europe/Paris"))
-    printContext = new DateTimePrintContext(zdt, Locale.ENGLISH, DecimalStyle.STANDARD)
-    parseContext = new DateTimeParseContext(Locale.ENGLISH, DecimalStyle.STANDARD, IsoChronology.INSTANCE)
+    printContext = new TTBPDateTimePrintContext(zdt, Locale.ENGLISH, DecimalStyle.STANDARD)
+    parseContext = new TTBPDateTimeParseContext(Locale.ENGLISH, DecimalStyle.STANDARD, IsoChronology.INSTANCE)
     buf = new java.lang.StringBuilder
   }
 }
