+++ shared/src/test/scala/org/threeten/bp/format/TestCharLiteralParser.scala
@@ -33,29 +33,30 @@ package org.threeten.bp.format
 
 import org.scalatest.FunSuite
 import org.threeten.bp.AssertionsHelper
+import org.threeten.bp.format.internal.TTBPDateTimeFormatterBuilder
 import org.threeten.bp.temporal.TemporalQueries
 
 /** Test CharLiteralPrinterParser. */
 class TestCharLiteralParser extends FunSuite with GenTestPrinterParser with AssertionsHelper {
   val data_success: List[List[Any]] = {
     List(
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "a", 0, 1),
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "aOTHER", 0, 1),
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "OTHERaOTHER", 5, 6),
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "OTHERa", 5, 6),
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "", 0, ~0),
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "a", 1, ~1),
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "A", 0, ~0),
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "b", 0, ~0),
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "OTHERbOTHER", 5, ~5),
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "OTHERb", 5, ~5),
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), false, "a", 0, 1),
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), false, "A", 0, 1))
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "a", 0, 1),
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "aOTHER", 0, 1),
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "OTHERaOTHER", 5, 6),
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "OTHERa", 5, 6),
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "", 0, ~0),
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "a", 1, ~1),
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "A", 0, ~0),
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "b", 0, ~0),
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "OTHERbOTHER", 5, ~5),
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), true, "OTHERb", 5, ~5),
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), false, "a", 0, 1),
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), false, "A", 0, 1))
   }
 
   test("parse_success") {
     data_success.foreach {
-      case (pp: DateTimeFormatterBuilder.CharLiteralPrinterParser) :: (caseSensitive: Boolean) :: (text: String) :: (pos: Int) :: (expectedPos: Int) :: Nil =>
+      case (pp: TTBPDateTimeFormatterBuilder.DateTimePrinterParser) :: (caseSensitive: Boolean) :: (text: String) :: (pos: Int) :: (expectedPos: Int) :: Nil =>
         parseContext.setCaseSensitive(caseSensitive)
         val result: Int = pp.parse(parseContext, text, pos)
         assertEquals(result, expectedPos)
@@ -68,13 +69,13 @@ class TestCharLiteralParser extends FunSuite with GenTestPrinterParser with Asse
 
   val data_error: List[List[Any]] = {
     List[List[Any]](
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), "a", -1, classOf[IndexOutOfBoundsException]),
-      List(new DateTimeFormatterBuilder.CharLiteralPrinterParser('a'), "a", 2, classOf[IndexOutOfBoundsException]))
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), "a", -1, classOf[IndexOutOfBoundsException]),
+      List(new TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser('a'), "a", 2, classOf[IndexOutOfBoundsException]))
   }
 
   test("parse_error") {
     data_error.foreach {
-      case (pp: DateTimeFormatterBuilder.CharLiteralPrinterParser) :: (text: String) :: (pos: Int) :: (expected: Class[_]) :: Nil =>
+      case (pp: TTBPDateTimeFormatterBuilder.CharLiteralPrinterParser) :: (text: String) :: (pos: Int) :: (expected: Class[_]) :: Nil =>
         try {
           pp.parse(parseContext, text, pos)
         }
